requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
anthropic
google-search-results==2.4.2 